+++ apps/web/components/simple-image/tailwind.tsx
@@ -5,6 +5,7 @@ export const component = (
   <Img
     alt="Ode Grinder"
     className="mx-auto"
+    width={250}
     height={250}
     src="/static/ode-grinder.jpg"
   />